Fixed #018627: Image not displayed on ez Image Editor extension when in eZDBFileHandler cluster mode.

+++ classes/image_analyzer.php
@@ -10,20 +10,26 @@
  */
 class eZIEImageAnalyzer extends ezcImageAnalyzer
 {
+    protected $deleteLocal = true;
+
     /**
      * Constructor overload
      * Creates a local copy of the image so that it can be analyzed
      *
      * @param string $file
      */
-    public function __construct( $file )
+    public function __construct( $file, $deleteLocal = true )
     {
         $clusterHandler = eZClusterFileHandler::instance( $file );
         $clusterHandler->fetch();
 
         parent::__construct( $file );
 
-        $clusterHandler->deleteLocal();
+        $this->deleteLocal = $deleteLocal;
+        if( $this->deleteLocal )
+        {
+            $clusterHandler->deleteLocal();
+        }
     }
 
     /**
@@ -39,7 +45,10 @@ class eZIEImageAnalyzer extends ezcImageAnalyzer
 
         parent::analyzeImage();
 
-        $clusterHandler->deleteLocal();
+        if( $this->deleteLocal )
+        {
+            $clusterHandler->deleteLocal();
+        }
     }
 }
